pyproject.toml
[project]
name = "resume-bot"
version = "0.1.0"
description = "A private RAG chatbot that answers questions about resume, LinkedIn and work experience — fast, factual, and source-backed."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.2",
    "google-generativeai>=0.8.5",
    "linkedin-scraper>=2.11.5",
    "openai>=1.108.0",
    "pdfplumber>=0.11.7",
    "pinecone>=7.3.0",
    "pydantic>=2.11.9",
    "pymilvus>=2.6.2",
    "pypandoc>=1.15",
    "pypdf2>=3.0.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "uvicorn>=0.35.0",
]
